consumer/requirements.txt
confluent-kafka==2.3.0
google-generativeai>=0.8.0  # 確保在 0.8.0 以上以支援 2.5 系列模型
psycopg2-binary